+++ Content.Shared/Armor/ComplexArmorComponent.cs
@@ -1,7 +1,8 @@
 using Content.Shared.Damage;
 using Content.Shared.Inventory;
 using Robust.Shared.GameStates;
-using Robust.Shared.Utility;
+using Robust.Shared.Prototypes;
+using Content.Shared.Tag;
 
 namespace Content.Shared.Armor;
 
@@ -12,14 +13,14 @@ namespace Content.Shared.Armor;
 public sealed partial class ComplexArmorComponent : Component, IClothingSlots
 {
     /// <summary>
-    ///     The damage reduction
+    ///     The damage reduction.
     /// </summary>
     [DataField(required: true), AutoNetworkedField]
     public DamageModifierSet Modifiers = default!;
 
-    [DataField, AutoNetworkedField]
+    [DataField(required: true), AutoNetworkedField]
     public SlotFlags Slots { get; set; } = SlotFlags.HEAD;
 
     [DataField(required: true), AutoNetworkedField]
-    public String ClothingTag = default!;
+    public ProtoId<TagPrototype>[] Tags = [];
 }
