+++ Resources/Prototypes/tags.yml
@@ -1048,6 +1048,9 @@
   id: Radio
 
 - type: Tag
+  id: RadiationProtectorHelmet
+
+- type: Tag
   id: RawMaterial
 
 # Give this to something that doesn't need any special recycler behavior and just needs deleting.
